pyproject.toml
[tool.poetry]
name = "lottery-app"
version = "0.1.0"
description = "Lottery System"
authors = ["João Paulo Ventorim <jp.ventorim@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.7"
django-ninja = "^1.2.2"
psycopg2-binary = "^2.9.9"
django-ninja-extra = "^0.21.2"
django-ninja-jwt = "^5.3.2"
pydantic = {extras = ["email"], version = "^2.8.2"}
celery = "^5.4.0"
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
ruff = "^0.5.6"
mypy = "^1.11.1"
coverage = "^7.6.1"
ipdb = "^0.13.13"
ipython = "^8.26.0"
django-stubs = "^5.0.4"
django-extensions = "^3.2.3"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
django_find_project = true
python_files = ["test*.py"]
addopts = [
    "--reuse-db",
    "--ds=lottery_app.settings",
    "--pdbcls=IPython.terminal.debugger:Pdb",
]
filterwarnings = [
    "ignore",
    "default:::lottery_app",
]

[tool.ruff]
line-length = 90
target-version = "py312"

[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "C90",  # mccabe (complexity)
    "COM",  # flake8-commas
    "DTZ",  # flake8-datetimez
    "E",    # pycodestyle
    "ERA",  # eradicate
    "F",    # pyflakes
    "FA",   # flake8-future-annotations
    "FLY",  # flynt
    "I",    # isort
    "INP",  # flake8-no-pep420
    "ISC",  # flake8-implicit-str-concat
    "LOG",  # flake8-logging
    "N",    # pep8-naming
    "PERF", # perflint
    "PIE",  # flake8-pie
    "PL",   # pylint
    "PTH",  # flake8-use-pathlib
    "Q",    # flake8-quotes
    "RET",  # flake8-return
    "RUF",  # Ruff-specific rules
    "S",    # flake8-bandit
    "SIM",  # flake8-simplify
    "T10",  # flake8-debugger
    "UP",   # pyupgrade
    "W",    # pycodestyle
]

[tool.ruff.lint.isort]
lines-after-imports = 2
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.isort.sections]
"django" = ['django']

[tool.ruff.lint.per-file-ignores]
"**/migrations/**" = ['RUF012', 'E501']
"**/test*/**" = ["S101","S105", "S106", "PLR2004"]

[tool.autopep8]
max_line_length = 90
aggressive = 3
experimental = true

[tool.mypy]
files = ["lottery_app"]
plugins = ["mypy_django_plugin.main", "pydantic.mypy"]
python_version = "3.12"
show_error_codes = true
exclude = ['conftest\.py']

disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["ninja_jwt.*", "celery.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "lottery_app.settings"